[expo-notifications] Move FirebaseListenerService to higher scope
This will let us reuse it for more than just token listening
